sendAfterSeconds유무에 따른 동작차이
